chainlit/requirements.txt
chainlit==2.0.dev0
pydantic==2.10.1
langchain>=0.0.300
psycopg2-binary==2.9.6
requests==2.31.0
langchain_community
langchain-ollama
langchain_openai
openai
chromadb
tiktoken
